backend/requirements.txt
Django==4.2.7
mysqlclient==2.1.1
graphene-django==3.0.0
django-cors-headers==4.3.0
django-graphql-jwt==0.3.4
gunicorn==21.2.0
gevent==23.9.1
langdetect==1.0.9
Pillow==10.0.1
python-dotenv==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
graphene-django-filter==0.6.5
#pymysql==1.0.3
